mcps/beeper/requirements.txt
fastmcp==3.0.2
python-dotenv>=1.0
beeper-desktop-api @ git+https://github.com/beeper/desktop-api-python.git
